+++ wiki/Changelog.md
@@ -1,3 +1,4 @@
+  * Tweaked DPI scaling to avoid terminal resizing on font selection (~#492).
   * Retaining config file comments (#574).
   * Enabling auto-scaling of window decoration (#547, #517, thanks to Jason Mansour #588).
   * Option AllowSetSelection to enable OSC 52 control sequence to set clipboard for pasting (#258, thanks to kangjianbin).